+++ CHANGES.rst
@@ -1,7 +1,9 @@
 0.1.2 (unreleased)
 ------------------
 
+- Rename package from kirlent.docutils to kirlent_docutils.
 - Change field markup for pauses (substep/fragment).
+- Restore support for Python 3.7 and 3.8.
 
 0.1.1 (2022-03-08)
 ------------------